% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFirstTwo.R
\name{addFirstTwo}
\alias{addFirstTwo}
\title{addFirstTwo}
\usage{
addFirstTwo(vec)
}
\arguments{
\item{vec}{: Must be a vector with numerical values. Exemple : x or c(2,5)}
}
\value{
Return the sum of the 2 first elements of the vector
}
\description{
Calculate the sume of the 2 first elements of a vector
}
\details{
# Arguments
}
\examples{
addFirstTwo(c(1,2))
addFirstTwo(c(3,4,5))

}
