% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RFM.R
\name{RFMfunction}
\alias{RFMfunction}
\title{Calculate RFM}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{: A data table containing the transaction recorded for every customer}

\item{weight_recency}{:   Weight of recency = a number}

\item{weight_frequency}{: Weight of frequency = a number}

\item{weight_monetary}{: Weight of monetary = a number}
}
\value{
a  data.data containing the recency, frequency and monetary
score as well as the weighted final score and the group membership.
}
\description{
Calculate the weighted RFM score : recency, frequency and monetary for every customer
}
\details{
# Arguments


\code{data} contains the transactional data. The dataset must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
